+++ tests/dom_test.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html></html>
+    <head></head>
+    <body data-demo-id="statemachine">
+        <div class="jtk-demo-main">
+            <div id="infobox">  
+            </div>
+            <div class="jtk-demo-canvas canvas-wide statemachine-demo jtk-surface jtk-surface-nopan" id="canvas">
+            </div>
+        </div>
+    </body>
+</html>
