+++ visualizer/src/css/index.css
@@ -1,3 +1,8 @@
+/* Copyright MIPTV 2019 
+ *
+ * @author Eric Konks exdevour@gmail.com
+*/
+
 .demo {
     /* for IE10+ touch devices */
     touch-action:none;
